+++ builds/4chan-XT.user.js
@@ -1,6 +1,6 @@
 // ==UserScript==
 // @name         4chan XT
-// @version      2.28.0
+// @version      2.28.1
 // @minGMVer     1.14
 // @minFFVer     115
 // @namespace    4chan-XT
@@ -158,8 +158,8 @@
   'use strict';
 
   var version = {
-    "version": "2.28.0",
-    "date": "2026-06-30T09:09:09Z"
+    "version": "2.28.1",
+    "date": "2026-07-01T09:09:09Z"
   }
   ;
 
@@ -1475,10 +1475,10 @@
       const iconName = [...el.classList].find(c => c in linkifyIcons);
       if (!iconName)
         return;
-      const span = document.createElement('span');
-      span.className = 'linkify-icon icon';
-      span.innerHTML = linkifyIcons[iconName];
-      el.prepend(span);
+      $.prepend(el, $.el('span', {
+        className: 'linkify-icon icon',
+        innerHTML: linkifyIcons[iconName],
+      }));
     },
     /** Get the raw SVG string for an icon. */
     get(name) {
@@ -7246,10 +7246,12 @@ svg.icon {
 .linkify-icon {
   position: absolute;
   left: 0;
-  top: 50%;
-  transform: translateY(-50%);
+  top: var(--linkify-icon-top, 0);
+  line-height: inherit;
   width: 14px;
-  height: 14px;
+  height: 1lh;
+  display: flex;
+  align-items: center;
   pointer-events: none;
 }
 .linkify.youtube .linkify-icon svg {
@@ -7534,7 +7536,7 @@ svg.icon {
 
   // cSpell:ignore installGentoo, webfont
   const mainCSS = style + variableBase + yotsuba + yotsubaB + futaba + burichan + tomorrow + photon + spooky + iconCss + fxTwitterCss;
-  const CSS = {
+  const CSS$1 = {
     boards: mainCSS,
     report,
     www,
@@ -7565,7 +7567,7 @@ svg.icon {
           return this.addStyle();
       },
       addStyle() {
-          return this.style = $.addStyle(CSS.sub(Conf.usercss), 'custom-css', '#fourchanx-css');
+          return this.style = $.addStyle(CSS$1.sub(Conf.usercss), 'custom-css', '#fourchanx-css');
       },
       rmStyle() {
           if (this.style) {
@@ -7577,7 +7579,7 @@ svg.icon {
           if (!this.style) {
               return this.addStyle();
           }
-          return this.style.textContent = CSS.sub(Conf.usercss);
+          return this.style.textContent = CSS$1.sub(Conf.usercss);
       }
   };
 
@@ -11266,7 +11268,7 @@ svg.icon {
         dateEl.dataset.fullTime = dateEl.textContent;
         dateEl.title = dateEl.textContent;
       }
-      return RelativeDates.update(this);
+      RelativeDates.update(this);
     },
     /** @param diff is milliseconds from now. */
     relative(diff, now, date, abbrev) {
@@ -11457,13 +11459,13 @@ svg.icon {
       node() {
           let a;
           if (a = $('.abbr > a:not([onclick])', this.nodes.comment)) {
-              return $.on(a, 'click', ExpandComment.cb);
+              $.on(a, 'click', ExpandComment.cb);
           }
       },
       callbacks: [],
       cb(e) {
           e.preventDefault();
-          return ExpandComment.expand(Get.postFromNode(this));
+          ExpandComment.expand(Get.postFromNode(this));
       },
       expand(post) {
           let a;
@@ -16730,6 +16732,8 @@ svg.icon {
           if (Conf['Comment Expansion']) {
               ExpandComment.callbacks.push(this.node);
           }
+          const isFirefox = CSS.supports('-moz-appearance', 'none');
+          doc.style.setProperty('--linkify-icon-top', isFirefox ? '-0.1em' : '0');
           Callbacks.Post.push({
               name: 'Linkify',
               cb: this.node
@@ -18391,7 +18395,7 @@ svg.icon {
                   node.removeAttribute('accesskey');
               }
           };
-          return $.on(d, '4chanXInitFinished', init);
+          $.on(d, '4chanXInitFinished', init);
       },
       sync(key, hotkey) {
           return Conf[hotkey] = key;
@@ -19201,7 +19205,7 @@ svg.icon {
           return $.ready(this.ready);
       },
       ready() {
-          $.addStyle(CSS.report);
+          $.addStyle(CSS$1.report);
           if (Conf['Archive Report']) {
               Report.archive();
           }
@@ -19615,7 +19619,7 @@ $\
       switch (location.hostname) {
         // case 'www.4chan.org': case 'www.4channel.org':
         case 'www.4chan.org':
-          $.onExists(doc, 'body', () => $.addStyle(CSS.www));
+          $.onExists(doc, 'body', () => $.addStyle(CSS$1.www));
           Captcha.replace.init();
           return;
         // case 'sys.4chan.org': case 'sys.4channel.org':
@@ -20464,7 +20468,7 @@ $\
       $.add(oldInfo, [this.file.link.previousSibling, this.file.link, this.file.link.nextSibling]);
       const info = $.el('span', { className: 'file-info' });
       FileInfo.format(Conf.fileInfo, this, info);
-      return $.prepend(this.file.text, info);
+      $.prepend(this.file.text, info);
     },
     format(formatString, post, outputNode) {
       let a;
@@ -21985,7 +21989,7 @@ Enable it on boards.${location.hostname.split('.')[1]}.org in your browser's pri
       toggle() {
           $.event('CloseMenu');
           $.set('Header catalog links', this.checked);
-          return CatalogLinks.set(this.checked);
+          CatalogLinks.set(this.checked);
       },
       set(useCatalog) {
           Conf['Header catalog links'] = useCatalog;
@@ -24375,7 +24379,7 @@ Enable it on boards.${location.hostname.split('.')[1]}.org in your browser's pri
               name: 'Disable Autoplaying Sounds',
               cb: this.node
           });
-          return $.ready(() => this.process(d.body));
+          $.ready(() => this.process(d.body));
       },
       stop(audio) {
           if (!audio.autoplay) {
@@ -24387,7 +24391,7 @@ Enable it on boards.${location.hostname.split('.')[1]}.org in your browser's pri
               return;
           }
           audio.controls = true;
-          return $.addClass(audio, 'controls-added');
+          $.addClass(audio, 'controls-added');
       },
       node() {
           return AntiAutoplay.process(this.nodes.comment);
@@ -24405,7 +24409,7 @@ Enable it on boards.${location.hostname.split('.')[1]}.org in your browser's pri
           if (window.getComputedStyle(el).display === 'none') {
               el.style.display = 'block';
           }
-          return $.addClass(el, 'autoplay-removed');
+          $.addClass(el, 'autoplay-removed');
       }
   };
 
@@ -24442,7 +24446,7 @@ Enable it on boards.${location.hostname.split('.')[1]}.org in your browser's pri
                   alt: '4chan',
                   src: '//s.4cdn.org/image/title/' + bannerCnt.dataset.src
               });
-              return $.add(bannerCnt, img);
+              $.add(bannerCnt, img);
           }
       },
       setTitle(title) {
@@ -24552,7 +24556,7 @@ Enable it on boards.${location.hostname.split('.')[1]}.org in your browser's pri
               return;
           }
           BoardConfig.ready(this.initBoard);
-          return PageReady.ready(this.initReady);
+          PageReady.ready(this.initReady);
       },
       initBoard() {
           if (g.BOARD.config.code_tags) {
@@ -26036,7 +26040,7 @@ Enable it on boards.${location.hostname.split('.')[1]}.org in your browser's pri
                   return $.toggleClass(doc, 'autohiding-scrollbar');
               }
           });
-          $.addStyle(CSS.sub(CSS.boards), 'fourchanx-css');
+          $.addStyle(CSS$1.sub(CSS$1.boards), 'fourchanx-css');
           Main.bgColorStyle = $.el('style', { id: 'fourchanx-bgcolor-css' });
           return Main.setClass();
       },